requirements.txt
Django==4.2
djangorestframework
django-cors-headers
sentence-transformers
chromadb
transformers
pypdf
python-docx
gunicorn
uvicorn
aiofiles
